backend/requirements.txt
# Requires Python 3.9–3.12 (not 3.14). Use: py -3.12 -m venv venv && venv\Scripts\activate
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
langchain==0.0.340
faiss-cpu>=1.7.4
pydantic>=2.5.0,<3
groq>=0.4.0
sentence-transformers>=2.2.0
openai>=1.3.0
langchain-openai==0.0.2